+++ lib/bullet/active_record60.rb
@@ -269,6 +269,17 @@ module Bullet
           end
         end
       )
+
+      ::ActiveRecord::Associations::CollectionProxy.prepend(
+        Module.new do
+          def count
+            if Bullet.start? && !proxy_association.is_a?(::ActiveRecord::Associations::ThroughAssociation)
+              Bullet::Detector::CounterCache.add_counter_cache(proxy_association.owner, proxy_association.reflection.name)
+            end
+            super
+          end
+        end
+      )
     end
   end
 end